Version 2.2.0
Added a special mode for Steam in-home Streaming, it fools Steam into
streaming the whole desktop then it brings the launched app to the
foreground.

Now, UWPHook will export the app icon to Steam when exporting new games,
it automatically decides on the biggest jpg to be used as icon.